backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
PyMuPDF
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
pydantic
